--- FlatShading/components/com_render.ts
@@ -1,9 +1,0 @@
-import {RenderDiffuse} from "./com_render_diffuse.js";
-import {RenderSpecular} from "./com_render_specular.js";
-
-export type Render = RenderDiffuse | RenderSpecular;
-
-export const enum RenderKind {
-    Diffuse,
-    Specular,
-}
