+++ FlatShading/world.ts
@@ -1,6 +1,6 @@
 import {Camera} from "./components/com_camera.js";
 import {Light} from "./components/com_light.js";
-import {Render} from "./components/com_render.js";
+import {Render} from "./components/com_render2.js";
 import {Rotate} from "./components/com_rotate.js";
 import {Transform} from "./components/com_transform.js";
