+++ core/world.ts
@@ -4,6 +4,7 @@ import {Camera} from "./components/com_camera.js";
 import {Draw} from "./components/com_draw.js";
 import {Light} from "./components/com_light.js";
 import {Move} from "./components/com_move.js";
+import {Rotate} from "./components/com_rotate.js";
 import {Transform} from "./components/com_transform.js";
 
 const enum Component {
@@ -14,6 +15,7 @@ const enum Component {
     Light,
     Move,
     Render,
+    Rotate,
     Transform,
 }
 
@@ -25,6 +27,7 @@ export const enum Has {
     Light = 1 << Component.Light,
     Move = 1 << Component.Move,
     Render = 1 << Component.Render,
+    Rotate = 1 << Component.Rotate,
     Transform = 1 << Component.Transform,
 }
 
@@ -40,5 +43,6 @@ export interface World {
     Light: Array<Light>;
     Move: Array<Move>;
     // Render depends on the version of WebGL. See com_render*, sys_render*.
+    Rotate: Array<Rotate>;
     Transform: Array<Transform>;
 }
